fixed clients compute_answer and remake output

+++ client.h
@@ -55,7 +55,17 @@ static void parse_args(int argc, char **argv, struct opts *options);
  */
 static uint8_t *read_from_server(int sockfd_con, uint8_t *buffer, size_t n);
 
-static uint8_t *send_to_server(int sockfd_con, uint8_t *buffer, size_t n);
+/**
+ * @brief Send message via socket
+ *
+ * This code *illustrates* one way to deal with partial reads
+ *
+ * @param sockfd_con Socket to send with
+ * @param buffer Buffer with data to write
+ * @param n Size to write
+ * @return Pointer to buffer on success, else NULL
+ */
+static uint16_t *send_to_server(int sockfd_con, uint16_t *buffer, size_t n);
 
 /**
  * @brief Compute answer to request
